+++ tmp/src/main/java/cn/liboshuai/scratch/tmp/RetryStrategy.java
@@ -3,7 +3,7 @@ package cn.liboshuai.scratch.tmp;
 import java.time.Duration;
 
 public interface RetryStrategy {
-    int getNumRemainingRetries();
+    long getNumRemainingRetries();
     Duration getCurrentRetryDelay();
     RetryStrategy getNextRetryStrategy();
 }
